# A more conservative fix for phase assembly performance.