Reverts the previous attempt to overhaul the algorithm which
proved buggy.

Prune the recursion when we re-visit a node via a shorter path from the
initial node.

Before last commit:

```
[info] # Run complete. Total time: 00:00:44
[info] Benchmark                        (size)  Mode  Cnt       Score      
Error  Units
[info] PhaseAssemblyBenchmark.assemble       1  avgt    5    1153.587 ±   
46.410  ns/op
[info] PhaseAssemblyBenchmark.assemble       4  avgt    5    4055.886 ±  
891.833  ns/op
[info] PhaseAssemblyBenchmark.assemble       8  avgt    5   19886.933 ±  
518.378  ns/op
[info] PhaseAssemblyBenchmark.assemble      16  avgt    5  768377.075 ±
28072.668  ns/op
```

After:

```
[info] # Run complete. Total time: 00:00:44
[info] Benchmark                        (size)  Mode  Cnt     Score    
Error  Units
[info] PhaseAssemblyBenchmark.assemble       1  avgt    5  1011.742 ±
164.448  ns/op
[info] PhaseAssemblyBenchmark.assemble       4  avgt    5  2485.863 ±
482.310  ns/op
[info] PhaseAssemblyBenchmark.assemble       8  avgt    5  4333.091 ±
199.798  ns/op
[info] PhaseAssemblyBenchmark.assemble      16  avgt    5  9644.064 ±
468.399  ns/op
```