@retronym 2019-09-03
I'll submit a backport to 2.12.x once this is merged.