@lrytz 2019-09-03
```
!!    1 - neg/t7622-cyclic-dependency               [output differs]
% diff /home/jenkins/workspace/scala-2.13.x-validate-main@2/test/files/neg/t7622-cyclic-dependency.check /home/jenkins/workspace/scala-2.13.x-validate-main@2/test/files/neg/t7622-cyclic-dependency-neg.log
@@ -1 +1 @@
-error: Cycle in phase dependencies detected at cyclicdependency1, created phase-cycle.dot
+error: Cycle in phase dependencies detected at cyclicdependency2, created phase-cycle.dot
```

probably just needs an update-check